+++ src/Vendr.PaymentProviders.Bambora/BamboraPaymentProviderBase.cs
@@ -1,0 +1,92 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using Vendr.Core;
+using Vendr.Core.Models;
+using Vendr.Core.Web.Api;
+using Vendr.Core.Web.PaymentProviders;
+using Vendr.PaymentProviders.Bambora.Api.Models;
+
+namespace Vendr.PaymentProviders.Bambora
+{
+    public abstract class BamboraPaymentProviderBase<TSettings> : PaymentProviderBase<TSettings>
+        where TSettings : BamboraSettingsBase, new()
+    {
+        public BamboraPaymentProviderBase(VendrContext vendr)
+            : base(vendr)
+        { }
+
+        public override string GetCancelUrl(OrderReadOnly order, TSettings settings)
+        {
+            settings.MustNotBeNull("settings");
+            settings.CancelUrl.MustNotBeNull("settings.CancelUrl");
+
+            return settings.CancelUrl;
+        }
+
+        public override string GetContinueUrl(OrderReadOnly order, TSettings settings)
+        {
+            settings.MustNotBeNull("settings");
+            settings.ContinueUrl.MustNotBeNull("settings.ContinueUrl");
+
+            return settings.ContinueUrl;
+        }
+
+        public override string GetErrorUrl(OrderReadOnly order, TSettings settings)
+        {
+            settings.MustNotBeNull("settings");
+            settings.ErrorUrl.MustNotBeNull("settings.ErrorUrl");
+
+            return settings.ErrorUrl;
+        }
+
+        protected BamboraClientConfig GetBamboraClientConfig(BamboraSettingsBase settings)
+        {
+            if (settings.TestMode)
+            {
+                return new BamboraClientConfig
+                {
+                    AccessKey = settings.TestAccessKey,
+                    MerchantNumber = settings.TestMerchantNumber,
+                    SecretKey = settings.TestSecretKey,
+                    MD5Key = settings.TestMd5Key
+                };
+            }
+            else
+            {
+                return new BamboraClientConfig
+                {
+                    AccessKey = settings.LiveAccessKey,
+                    MerchantNumber = settings.LiveMerchantNumber,
+                    SecretKey = settings.LiveSecretKey,
+                    MD5Key = settings.LiveMd5Key
+                };
+            }
+        }
+
+        protected PaymentStatus GetPaymentStatus(BamboraTransaction transaction)
+        {
+            if (transaction.Total.Credited > 0)
+                return PaymentStatus.Refunded;
+
+            if (transaction.Total.Declined > 0)
+                return PaymentStatus.Cancelled;
+
+            if (transaction.Total.Captured > 0)
+                return PaymentStatus.Captured;
+
+            if (transaction.Total.Authorized > 0)
+                return PaymentStatus.Authorized;
+
+            return PaymentStatus.Initialized;
+        }
+
+        protected string BamboraSafeOrderId(string orderId)
+        {
+            return Regex.Replace(orderId, "[^a-zA-Z0-9]", "");
+        }
+    }
+}
