+++ src/Vendr.PaymentProviders.Bambora/BamboraSettingsBase.cs
@@ -2,7 +2,7 @@ using Vendr.Core.Web.PaymentProviders;
 
 namespace Vendr.PaymentProviders.Bambora
 {
-    public class BamboraSettings
+    public class BamboraSettingsBase
     {
         [PaymentProviderSetting(Name = "Continue URL",
             Description = "The URL to continue to after this provider has done processing. eg: /continue/",
