+++ src/Vendr.PaymentProviders.Bambora/Vendr.PaymentProviders.Bambora.csproj
@@ -105,8 +105,10 @@
     <Compile Include="Api\Models\BamboraResponseMetaData.cs" />
     <Compile Include="Api\Models\BamboraTransaction.cs" />
     <Compile Include="Api\Models\BamboraTransactionResponse.cs" />
-    <Compile Include="BamboraPaymentProvider.cs" />
-    <Compile Include="BamboraSettings.cs" />
+    <Compile Include="BamboraCheckoutOneTimeSettings.cs" />
+    <Compile Include="BamboraCheckoutOneTimePaymentProvider.cs" />
+    <Compile Include="BamboraPaymentProviderBase.cs" />
+    <Compile Include="BamboraSettingsBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\VersionInfo.cs" />
   </ItemGroup>
